s_azure/sbus/requirements.txt
azure-servicebus==0.50.2